Kamila

+++ SIRekrutmenH9/src/main/java/apap/tk/SIRekrutmenH9/service/LowonganServiceImpl.java
@@ -20,7 +20,7 @@ public class LowonganServiceImpl implements LowonganService{
     }
 
     @Override
-    public LowonganModel getLowonganById(Integer id){
+    public LowonganModel getLowonganById(Long id){
         return lowonganDB.findById(id).get();
     }
